refactoring adding base url using async
